Build shipping method
